Core/requirements.txt
-e git+git@bitbucket.org:almostthere/almostthere-Core.git@417764e2b8a72c35cdd5f41d51b505a10da7bbca#egg=AlmostThere_Core-develop
blist==1.3.6
boto==2.32.1
cassandra-driver==2.1.0
coverage==3.7.1
futures==2.1.6
nose==1.3.4
pymongo==2.7.2
redis==2.10.3
six==1.7.3
